# public/resources/api/api_docs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "SkyCrypt API Documentation",
        "version": "hmmm",
        "description": "Welcome to the SkyCrypt API Documentation page!<br> Here you can find information on all the endpoints we have, and how to use them.<br><br> The main API URL is `https://sky.shiiyu.moe/api/v2/`",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "https://sky.shiiyu.moe/api/v2"
        }
    ],
...
    "paths": {
        "/profile/${playerName}": {
            "get": {
                "tags": ["profile"],
                "summary": "Get a list of a player's profiles, with extensive data.",
                "parameters": [
                    {
                        "name": "playerName",
                        "in": "path",
                        "description": "Username of the player.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "PROFILE_ID": {
                                            "type": "object",
                                            "properties": {
                                                "profile_id": {
                                                    "type": "string",
                                                    "description": "The profile's unique ID."
                                                },
                                                "cute_name": {
                                                    "type": "string",
                                                    "description": "Profile's cute name, e.g. Lime, Strawberry"
                                                },
                                                "current": {
                                                    "type": "boolean",
                                                    "description": "If the profile was the most saved."
                                                },
                                                "last_save": {
                                                    "type": "integer",
                                                    "description": "The last time the profile was saved as a unix epoch."
                                                },
                                                "raw": {
                                                    "type": "object",
                                                    "description": "Raw Hypixel API data."
                                                },
                                                "items": {
                                                    "type": "object",
                                                    "description": "Items in a player's inventory, bags, or enderchest."
                                                },
                                                "data": {
                                                    "type": "object",
                                                    "description": "All other profile data, such as skills, slayers, stats, and more."
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/talismans/{playerName}": {
            "get": {
                "tags": ["talismans"],
                "summary": "Get a list of a player's profiles with talisman data.",
                "parameters": [
                    {
                        "name": "playerName",
                        "in": "path",
                        "description": "Username of the player.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Talismans"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/talismans/{playerName}/{profileName}": {
            "get": {
                "tags": ["talismans"],
                "summary": "Get a player's profile with talisman data.",
                "parameters": [
                    {
                        "name": "playerName",
                        "in": "path",
                        "description": "Username of the player.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "profileName",
                        "in": "path",
                        "description": "Profile cute name.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "profiles": {
                                            "type": "object",
                                            "properties": {
                                                "PROFILE_ID": {
                                                    "$ref": "#/components/schemas/Talismans"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/slayers/{playerName}": {
            "get": {
                "tags": ["slayers"],
                "summary": "Get a list of a player's profiles with slayer data.",
                "parameters": [
                    {
                        "name": "playerName",
                        "in": "path",
                        "description": "Username of the player.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Slayer"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/slayers/{playerName}/{profileName}": {
            "get": {
                "tags": ["slayers"],
                "summary": "Get a player's profile with slayer data.",
                "parameters": [
                    {
                        "name": "playerName",
                        "in": "path",
                        "description": "Username of the player.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "profileName",
                        "in": "path",
                        "description": "Profile cute name.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "profiles": {
                                            "type": "object",
                                            "properties": {
                                                "PROFILE_ID": {
                                                    "$ref": "#/components/schemas/Slayer"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/coins/{playerName}": {
            "get": {
                "tags": ["coins"],
                "summary": "Get a list of a player's profiles with bank and purse data.",
                "parameters": [
                    {
                        "name": "playerName",
                        "in": "path",
                        "description": "Username of the player.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Coins"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/coins/{playerName}/{profileName}": {
            "get": {
                "tags": ["coins"],
                "summary": "Get a player's profile with bank and purse data.",
                "parameters": [
                    {
                        "name": "playerName",
                        "in": "path",
                        "description": "Username of the player.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "profileName",
                        "in": "path",
                        "description": "Profile cute name.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "profiles": {
                                            "type": "object",
                                            "properties": {
                                                "PROFILE_ID": {
                                                    "$ref": "#/components/schemas/Coins"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/dungeons/{playerName}": {
            "get": {
                "tags": ["dungeons"],
                "summary": "Get a list of a player's profiles with dungeon and class data.",
                "parameters": [
                    {
                        "name": "playerName",
                        "in": "path",
                        "description": "Username of the player.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "profiles": {
                                            "type": "object",
                                            "properties": {
                                                "PROFILE_ID": {
                                                    "$ref": "#/components/schemas/Dungeons"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/dungeons/{playerName}/{profileName}": {
            "get": {
                "tags": ["dungeons"],
                "summary": "Get a player's profile with dungeon and class data.",
                "parameters": [
                    {
                        "name": "playerName",
                        "in": "path",
                        "description": "Username of the player.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "profileName",
                        "in": "path",
                        "description": "Profile cute name.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Dungeons"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Coins": {
                "type": "object",
                "properties": {
                    "profile_id": {
                        "type": "string",
                        "description": "The profile's unique ID."
                    },
                    "cute_name": {
                        "type": "string",
                        "description": "Profile's cute name, e.g. Lime, Strawberry"
                    },
                    "purse": {
                        "type": "integer"
                    },
                    "bank": {
                        "type": "integer"
                    }
                },
                "xml": {
                    "name": "Coins"
                }
            },
...
        }
    }
}